migrate: move vetCenterOutstation to products directory
⚠️ NOTE: This template is not yet wired up in [[...slug]].tsx routing

- Migrated vetCenterOutstation product files to src/products/vetCenterOutstation/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved query.ts from src/data/queries/ to src/products/vetCenterOutstation/
- Manually moved test file and renamed vetCenterOutstations.test.ts -> query.test.ts
- Moved formatted type from src/types/formatted/ to src/products/vetCenterOutstation/formatted-type.ts
- Moved both mock files: vetCenterOutstation.mock.json -> mock.json and vetCenterOutstation.mock.tsx -> mock.tsx
- Manually moved test snapshots to src/products/vetCenterOutstation/__snapshots__/
- Updated all import statements across the codebase
- Fixed import paths in migrated files (formatted-type.ts, query.ts, template.test.tsx, query.test.ts)
- All TypeScript errors resolved
- All tests passing (3 query tests, 1 template test)
- Updated migration guide with completed status and routing caveat
- Added CODEOWNERS entry for new product directory

IMPORTANT: vetCenterOutstation pages cannot be browser-tested as the template
is not connected to routing in [[...slug]].tsx. The template exists but was
never completed/wired up. Migration is structurally complete.
