+++ .github/CODEOWNERS
@@ -23,3 +23,4 @@ src/products/leadershipListing/ @department-of-veterans-affairs/ap-team
 src/products/vamcSystemVaPolice/ @department-of-veterans-affairs/ap-team
 src/products/vetCenter/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
 src/products/vetCenterLocationListing/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
+src/products/vetCenterOutstation/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
