+++ docs/ai-product-migration-guide.md
@@ -141,6 +141,7 @@ The `move-product.sh` script performs these actions:
 - [x] leadershipListing
 - [x] vetCenter
 - [x] vetCenterLocationListing
+- [x] vetCenterOutstation ⚠️
 
 ### 🔄 Pending Migration
 
@@ -154,28 +155,27 @@ The `move-product.sh` script performs these actions:
 - [ ] **vamcSystem** - Has template layout and query file
 - [ ] **vbaFacility** - Has template layout and query file
 
-- [ ] **vetCenterOutstation** - Has template layout and query file
-
 #### Priority Order Recommendation
 
-1. **vetCenterOutstation** - Related to vetCenter
-2. **vamcSystem** - Core VAMC functionality
-3. **healthCareLocalFacility** - Healthcare facility core
-4. **vamcHealthServicesListing** - Related to VAMC
-5. **locationsListing** - General locations
-6. **vbaFacility** - VBA facilities
-7. **resourcesSupport** - Support resources
-8. **questionAnswer** - Q&A functionality
+1. **vamcSystem** - Core VAMC functionality
+2. **healthCareLocalFacility** - Healthcare facility core
+3. **vamcHealthServicesListing** - Related to VAMC
+4. **locationsListing** - General locations
+5. **vbaFacility** - VBA facilities
+6. **resourcesSupport** - Support resources
+7. **questionAnswer** - Q&A functionality
 
 ### 📝 Migration Progress Tracking
 
-**Current Product**: vetCenterOutstation _(next up)_
+**Current Product**: vamcSystem _(next up)_
 
 **Status**: Ready for next migration
 
-**Last Completed**: vetCenterLocationListing - Successfully migrated with all tests passing
+**Last Completed**: vetCenterOutstation - Successfully migrated with all tests passing ⚠️
 
-**Notes**: vetCenterLocationListing migration completed successfully. Fixed import paths in formatted-type.ts and template.test.tsx. All TypeScript errors resolved and all tests passing (7 query tests + 12 template tests + 17 component tests).
+**Notes**: vetCenterOutstation migration completed successfully. Fixed import paths in formatted-type.ts, query.ts, template.test.tsx, and query.test.ts. Manually moved test files and mock files that weren't caught by script. All TypeScript errors resolved and all tests passing (3 query tests + 1 template test).
+
+⚠️ **IMPORTANT**: vetCenterOutstation template is not yet wired up in [[...slug]].tsx routing - the template exists but has no route. This is expected as the template was never completed. Migration successful but pages cannot be accessed via browser until routing is implemented.
 
 ---
