+++ src/data/queries/index.ts
@@ -49,7 +49,7 @@ import * as VamcSystem from '@/data/queries/vamcSystem'
 import * as VamcSystemVaPolice from '@/products/vamcSystemVaPolice/query'
 import * as VetCenter from '@/products/vetCenter/query'
 import * as VetCenterHealthServices from '@/data/queries/vetCenterHealthServices'
-import * as VetCenterOutstation from '@/data/queries/vetCenterOutstation'
+import * as VetCenterOutstation from '@/products/vetCenterOutstation/query'
 import * as VetCenterLocationListing from '@/products/vetCenterLocationListing/query'
 import * as Wysiwyg from '@/data/queries/wysiwyg'
 import * as LeadershipListing from '@/products/leadershipListing/query'
