+++ src/products/vetCenterOutstation/template.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
-import { VetCenterOutstation } from './index'
-import { mockResponse } from '@/mocks/vetCenterOutstation.mock'
+import { VetCenterOutstation } from './template'
+import { mockResponse } from '@/products/vetCenterOutstation/mock'
 
 describe('VetCenterOutstation with valid data', () => {
   test('renders VetCenterOutstation component', () => {
